+++ src/audio_format.c
@@ -52,7 +52,7 @@ mpd_parse_audio_format(struct mpd_audio_format *audio_format, const char *p)
 			audio_format->bits = MPD_SAMPLE_FORMAT_DSD;
 
 			p = endptr + 1;
-			audio_format->channels = strtoul(p, NULL, 10);
+			audio_format->channels = (uint8_t)strtoul(p, NULL, 10);
 			return;
 		}
 	}
@@ -69,12 +69,12 @@ mpd_parse_audio_format(struct mpd_audio_format *audio_format, const char *p)
 			audio_format->bits = MPD_SAMPLE_FORMAT_DSD;
 			p += 4;
 		} else {
-			audio_format->bits = strtoul(p, &endptr, 10);
+			audio_format->bits = (uint8_t)strtoul(p, &endptr, 10);
 			p = *endptr == ':' ? endptr + 1 : NULL;
 		}
 
 		audio_format->channels = p != NULL
-			? strtoul(p, NULL, 10)
+			? (uint8_t)strtoul(p, NULL, 10)
 			: 0;
 	} else {
 		audio_format->bits = 0;
